+++ gdal/src/test/scala/geotrellis/contrib/vlm/spark/GDALRasterSourceRDDSpec.scala
@@ -61,8 +61,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
 
       info(s"RasterSource CRS: ${reprojectedSource.crs}")
 
-      val rdd = RasterSourceRDD(reprojectedSource, layout)
-
+      val rdd = RasterSourceRDD.spatial(reprojectedSource, layout)
 
       val actualKeys = rdd.keys.collect().sortBy { key => (key.col, key.row) }
 
@@ -73,7 +72,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
 
     it("should read in the tiles as squares") {
       val reprojectedRasterSource = rasterSource.reprojectToGrid(targetCRS, layout)
-      val rdd = RasterSourceRDD(reprojectedRasterSource, layout)
+      val rdd = RasterSourceRDD.spatial(reprojectedRasterSource, layout)
       val rows = rdd.collect()
 
       forAll(rows) { case (key, tile) =>
@@ -93,14 +92,8 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
     val geoTiffRDD = HadoopGeoTiffRDD.spatialMultiband(uri)
     val md = geoTiffRDD.collectMetadata[SpatialKey](floatingLayout)._2
 
-    val reprojectedExpectedRDD: MultibandTileLayerRDD[SpatialKey] = {
-      geoTiffRDD
-        .tileToLayout(md)
-        .reproject(
-          targetCRS,
-          layout
-        )._2.persist()
-    }
+    val reprojectedExpectedRDD: MultibandTileLayerRDD[SpatialKey] =
+      geoTiffRDD.tileToLayout(md).reproject(targetCRS, layout)._2.persist()
 
     def assertRDDLayersEqual(
       expected: MultibandTileLayerRDD[SpatialKey],
@@ -139,7 +132,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
 
     it("should reproduce tileToLayout") {
       // This should be the same as result of .tileToLayout(md.layout)
-      val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(rasterSource, md.layout)
+      val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD.spatial(rasterSource, md.layout)
 
       // Complete the reprojection
       val reprojectedSource = rasterSourceRDD.reproject(targetCRS, layout)._2
@@ -149,7 +142,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
 
     it("should reproduce tileToLayout followed by reproject") {
       val reprojectedSourceRDD: MultibandTileLayerRDD[SpatialKey] =
-        RasterSourceRDD(rasterSource.reprojectToGrid(targetCRS, layout), layout)
+        RasterSourceRDD.spatial(rasterSource.reprojectToGrid(targetCRS, layout), layout)
 
       // geotrellis.raster.io.geotiff.GeoTiff(reprojectedExpectedRDD.stitch, targetCRS).write("/tmp/expected.tif")
       // geotrellis.raster.io.geotiff.GeoTiff(reprojectedSourceRDD.stitch, targetCRS).write("/tmp/actual.tif")
@@ -193,7 +186,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
         val rasterSource = GDALRasterSource(uri)
 
         // This should be the same as result of .tileToLayout(md.layout)
-        val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(rasterSource, md.layout)
+        val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD.spatial(rasterSource, md.layout)
         // Complete the reprojection
         val reprojectedSource = rasterSourceRDD.reproject(targetCRS, layout)._2
 
@@ -202,12 +195,12 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
 
       it("should reproduce tileToLayout followed by reproject GDAL") {
         val expectedRasterSource = GDALRasterSource(expectedFilePath)
-        val reprojectedExpectedRDDGDAL: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(expectedRasterSource, layout)
+        val reprojectedExpectedRDDGDAL: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD.spatial(expectedRasterSource, layout)
         val rasterSource = GDALRasterSource(uri)
         val reprojectedRasterSource = rasterSource.reprojectToGrid(targetCRS, layout)
 
         // This should be the same as .tileToLayout(md.layout).reproject(crs, layout)
-        val reprojectedSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(reprojectedRasterSource, layout)
+        val reprojectedSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD.spatial(reprojectedRasterSource, layout)
 
         assertRDDLayersEqual(reprojectedExpectedRDDGDAL, reprojectedSourceRDD, true)
       }
@@ -216,7 +209,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
         println(java.lang.Thread.activeCount())
 
         /** Functions to trigger Datasets computation */
-        def ltsWithDatasetsTriggered(lts: LayoutTileSource): LayoutTileSource = { rsWithDatasetsTriggered(lts.source); lts }
+        def ltsWithDatasetsTriggered(lts: LayoutTileSource[SpatialKey]): LayoutTileSource[SpatialKey] = { rsWithDatasetsTriggered(lts.source); lts }
         def rsWithDatasetsTriggered(rs: RasterSource): RasterSource = {
           val brs = rs.asInstanceOf[GDALRasterSource]
           brs.dataset.rasterExtent
@@ -225,7 +218,7 @@ class GDALRasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRa
         }
 
         /** Do smth usual with the original RasterSource to force VRTs allocation */
-        def reprojRS(i: Int): LayoutTileSource =
+        def reprojRS(i: Int): LayoutTileSource[SpatialKey] =
           ltsWithDatasetsTriggered(
             rsWithDatasetsTriggered(
               rsWithDatasetsTriggered(GDALRasterSource(filePathByIndex(i)))