+++ vlm/src/test/scala/geotrellis/contrib/vlm/spark/RasterSourceRDDSpec.scala
@@ -56,8 +56,7 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
       info(s"RasterSource CRS: ${reprojectedSource.crs}")
 
-      val rdd = RasterSourceRDD(reprojectedSource, layout)
-
+      val rdd = RasterSourceRDD.spatial(reprojectedSource, layout)
 
       val actualKeys = rdd.keys.collect().sortBy { key => (key.col, key.row) }
 
@@ -68,7 +67,7 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
     it("should read in the tiles as squares") {
       val reprojectedRasterSource = rasterSource.reprojectToGrid(targetCRS, layout)
-      val rdd = RasterSourceRDD(reprojectedRasterSource, layout)
+      val rdd = RasterSourceRDD.spatial(reprojectedRasterSource, layout)
       val rows = rdd.collect()
 
       forAll(rows) { case (key, tile) =>
@@ -98,10 +97,10 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
     }
 
     def assertRDDLayersEqual(
-                              expected: MultibandTileLayerRDD[SpatialKey],
-                              actual: MultibandTileLayerRDD[SpatialKey],
-                              matchRasters: Boolean = false
-                            ): Unit = {
+      expected: MultibandTileLayerRDD[SpatialKey],
+      actual: MultibandTileLayerRDD[SpatialKey],
+      matchRasters: Boolean = false
+    ): Unit = {
       val joinedRDD = expected.filter { case (_, t) => !t.band(0).isNoDataTile }.leftOuterJoin(actual)
 
       joinedRDD.collect().foreach { case (key, (expected, actualTile)) =>
@@ -134,7 +133,7 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
     it("should reproduce tileToLayout") {
       // This should be the same as result of .tileToLayout(md.layout)
-      val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(rasterSource, md.layout)
+      val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD.spatial(rasterSource, md.layout)
 
       // Complete the reprojection
       val reprojectedSource = rasterSourceRDD.reproject(targetCRS, layout)._2
@@ -144,7 +143,7 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
     it("should reproduce tileToLayout followed by reproject") {
       val reprojectedSourceRDD: MultibandTileLayerRDD[SpatialKey] =
-        RasterSourceRDD(rasterSource.reprojectToGrid(targetCRS, layout), layout)
+        RasterSourceRDD.spatial(rasterSource.reprojectToGrid(targetCRS, layout), layout)
 
       // geotrellis.raster.io.geotiff.GeoTiff(reprojectedExpectedRDD.stitch, targetCRS).write("/tmp/expected.tif")
       // geotrellis.raster.io.geotiff.GeoTiff(reprojectedSourceRDD.stitch, targetCRS).write("/tmp/actual.tif")