+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Expose GDAL Metadata API, expose RasterSources Metadata API 
+- LayoutTileSource[K] implementation
 
 ## [3.17.0] - 2019-07-23